+++ soundscapes/app/frontend/SoundScapesFrontend/actions/filter_actions.js
@@ -1,0 +1,14 @@
+import { getSoundscapes } from './soundscape_actions';
+
+export const UPDATE_FILTER = 'UPDATE_FILTER';
+
+export const changeFilter = (filter, value) => ({
+  type: UPDATE_FILTER,
+  filter,
+  value
+});
+
+export const updateFilter = (filter, value) => (dispatch, getState) => {
+  dispatch(changeFilter(filter, value));
+  return getSoundscapes(getState().ui.filters)(dispatch);
+};
