+++ soundscapes/app/frontend/SoundScapesFrontend/reducers/filters_reducer.js
@@ -1,14 +1,1 @@
-import { getSoundscapes } from './soundscape_actions';
 
-export const UPDATE_FILTER = 'UPDATE_FILTER';
-
-export const changeFilter = (filter, value) => ({
-  type: UPDATE_FILTER,
-  filter,
-  value
-});
-
-export const updateFilter = (filter, value) => (dispatch, getState) => {
-  dispatch(changeFilter(filter, value));
-  return getSoundscapes(getState().ui.filters)(dispatch);
-};
